Changing a lot of NSLogs to SysPrintf on OS X's PCSXR. Enable the emulog on OS X debug, and have create and use a log in ~/Library/Logs/PCSXR. Delete the hack that enabled PCSXR logs to go to the console in debug mode. Use modern syntax to create static NSArrays (and retain them in non-ARC code). Simplify the pauseSafeWithBlock to just create a new dispatch queue and run pauseSafe on the queue.
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@85931 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ macosx/plugins/DFXVideo/macsrc/PluginGLView.m
@@ -695,7 +695,7 @@ void BlitScreen16NS(unsigned char * surf,long x,long y)
 			}
 			else
 			{
-				int LineOffset,SurfOffset;
+				long LineOffset,SurfOffset;
 				GLuint * SRCPtr = (GLuint *)(psxVuw + (y << 10) + x);
 				GLuint * DSTPtr =
 				((GLuint *)surf) + (PreviousPSXDisplay.Range.x0 >> 1);
